Move internal classes to new internal model and registry package
