+++ org.eclipse.tm4e.languageconfiguration/src/main/java/org/eclipse/tm4e/languageconfiguration/internal/model/EnterActionAndIndent.java
@@ -9,7 +9,7 @@
  * Contributors:
  * Angelo Zerr <angelo.zerr@gmail.com> - initial API and implementation
  */
-package org.eclipse.tm4e.languageconfiguration.internal.supports;
+package org.eclipse.tm4e.languageconfiguration.internal.model;
 
 public final class EnterActionAndIndent {
