+++ org.eclipse.tm4e.languageconfiguration/src/main/java/org/eclipse/tm4e/languageconfiguration/internal/model/ILanguageConfiguration.java
@@ -9,16 +9,11 @@
  * Contributors:
  * Lucas Bullen (Red Hat Inc.) - initial API and implementation
  */
-package org.eclipse.tm4e.languageconfiguration.internal;
+package org.eclipse.tm4e.languageconfiguration.internal.model;
 
 import java.util.List;
 
 import org.eclipse.jdt.annotation.Nullable;
-import org.eclipse.tm4e.languageconfiguration.internal.supports.StandardAutoClosingPairConditional;
-import org.eclipse.tm4e.languageconfiguration.internal.supports.CharacterPair;
-import org.eclipse.tm4e.languageconfiguration.internal.supports.CommentRule;
-import org.eclipse.tm4e.languageconfiguration.internal.supports.FoldingRule;
-import org.eclipse.tm4e.languageconfiguration.internal.supports.OnEnterRule;
 
 /**
  * @see <a href=
