+++ org.eclipse.tm4e.languageconfiguration/src/main/java/org/eclipse/tm4e/languageconfiguration/internal/model/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Types in this package are modeled after <a href=
+ * "https://github.com/microsoft/vscode/blob/main/src/vs/editor/common/languages/languageConfiguration.ts">
+ * github.com/microsoft/vscode/blob/main/src/vs/editor/common/languages/languageConfiguration.ts</a>
+ */
+@NonNullByDefault
+package org.eclipse.tm4e.languageconfiguration.internal.model;
+
+import org.eclipse.jdt.annotation.NonNullByDefault;
