+++ org.eclipse.tm4e.languageconfiguration/src/main/java/org/eclipse/tm4e/languageconfiguration/internal/preferences/LanguageConfigurationLabelProvider.java
@@ -15,7 +15,7 @@ import org.eclipse.jdt.annotation.Nullable;
 import org.eclipse.jface.viewers.ITableLabelProvider;
 import org.eclipse.jface.viewers.LabelProvider;
 import org.eclipse.swt.graphics.Image;
-import org.eclipse.tm4e.languageconfiguration.internal.ILanguageConfigurationDefinition;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.ILanguageConfigurationDefinition;
 
 final class LanguageConfigurationLabelProvider extends LabelProvider implements ITableLabelProvider {
