+++ org.eclipse.tm4e.languageconfiguration/src/main/java/org/eclipse/tm4e/languageconfiguration/internal/preferences/LanguageConfigurationPreferencePage.java
@@ -43,10 +43,10 @@ import org.eclipse.swt.widgets.Label;
 import org.eclipse.swt.widgets.Listener;
 import org.eclipse.swt.widgets.Table;
 import org.eclipse.swt.widgets.TableColumn;
-import org.eclipse.tm4e.languageconfiguration.internal.ILanguageConfigurationDefinition;
-import org.eclipse.tm4e.languageconfiguration.internal.ILanguageConfigurationRegistryManager;
-import org.eclipse.tm4e.languageconfiguration.internal.LanguageConfigurationRegistryManager;
-import org.eclipse.tm4e.languageconfiguration.internal.WorkingCopyLanguageConfigurationRegistryManager;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.ILanguageConfigurationDefinition;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.ILanguageConfigurationRegistryManager;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.LanguageConfigurationRegistryManager;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.WorkingCopyLanguageConfigurationRegistryManager;
 import org.eclipse.tm4e.languageconfiguration.internal.widgets.ColumnSelectionAdapter;
 import org.eclipse.tm4e.languageconfiguration.internal.widgets.ColumnViewerComparator;
 import org.eclipse.tm4e.languageconfiguration.internal.widgets.LanguageConfigurationPreferencesWidget;
