+++ org.eclipse.tm4e.languageconfiguration/src/main/java/org/eclipse/tm4e/languageconfiguration/internal/registry/AbstractLanguageConfigurationRegistryManager.java
@@ -9,7 +9,7 @@
  *  Contributors:
  *  Lucas Bullen (Red Hat Inc.) - initial API and implementation
  */
-package org.eclipse.tm4e.languageconfiguration.internal;
+package org.eclipse.tm4e.languageconfiguration.internal.registry;
 
 import java.util.HashMap;
 import java.util.HashSet;
@@ -17,6 +17,7 @@ import java.util.Map;
 
 import org.eclipse.core.runtime.content.IContentType;
 import org.eclipse.jdt.annotation.Nullable;
+import org.eclipse.tm4e.languageconfiguration.internal.model.ILanguageConfiguration;
 
 public abstract class AbstractLanguageConfigurationRegistryManager implements ILanguageConfigurationRegistryManager {
