+++ org.eclipse.tm4e.languageconfiguration/src/main/java/org/eclipse/tm4e/languageconfiguration/internal/registry/package-info.java
@@ -1,0 +1,4 @@
+@NonNullByDefault
+package org.eclipse.tm4e.languageconfiguration.internal.registry;
+
+import org.eclipse.jdt.annotation.NonNullByDefault;
