+++ org.eclipse.tm4e.languageconfiguration/src/main/java/org/eclipse/tm4e/languageconfiguration/internal/wizards/LanguageConfigurationImportWizard.java
@@ -14,9 +14,9 @@ package org.eclipse.tm4e.languageconfiguration.internal.wizards;
 import org.eclipse.jdt.annotation.Nullable;
 import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.jface.wizard.Wizard;
-import org.eclipse.tm4e.languageconfiguration.internal.ILanguageConfigurationDefinition;
-import org.eclipse.tm4e.languageconfiguration.internal.ILanguageConfigurationRegistryManager;
-import org.eclipse.tm4e.languageconfiguration.internal.LanguageConfigurationRegistryManager;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.ILanguageConfigurationDefinition;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.ILanguageConfigurationRegistryManager;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.LanguageConfigurationRegistryManager;
 import org.eclipse.ui.IImportWizard;
 import org.eclipse.ui.IWorkbench;
 import org.osgi.service.prefs.BackingStoreException;
