+++ org.eclipse.tm4e.languageconfiguration/src/main/java/org/eclipse/tm4e/languageconfiguration/internal/wizards/SelectLanguageConfigurationWizardPage.java
@@ -47,10 +47,10 @@ import org.eclipse.swt.widgets.Label;
 import org.eclipse.swt.widgets.Listener;
 import org.eclipse.swt.widgets.Text;
 import org.eclipse.tm4e.languageconfiguration.LanguageConfigurationPlugin;
-import org.eclipse.tm4e.languageconfiguration.internal.ILanguageConfigurationDefinition;
-import org.eclipse.tm4e.languageconfiguration.internal.ILanguageConfigurationRegistryManager;
-import org.eclipse.tm4e.languageconfiguration.internal.LanguageConfiguration;
-import org.eclipse.tm4e.languageconfiguration.internal.LanguageConfigurationDefinition;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.ILanguageConfigurationDefinition;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.ILanguageConfigurationRegistryManager;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.LanguageConfiguration;
+import org.eclipse.tm4e.languageconfiguration.internal.registry.LanguageConfigurationDefinition;
 import org.eclipse.tm4e.languageconfiguration.internal.widgets.LanguageConfigurationInfoWidget;
 import org.eclipse.tm4e.ui.internal.utils.ContentTypeHelper;
 import org.eclipse.ui.dialogs.ResourceSelectionDialog;
